fix: Fix scripts

+++ scripts/build.sh
@@ -1,5 +1,7 @@
 docker compose up -d
 
-#save new image to file tar
-rm uniqlo-server-prod.tar
+if [ -f uniqlo-server-prod.tar ]; then
+    rm uniqlo-server-prod.tar
+fi
+
 docker save uniqlo-server-prod > uniqlo-server-prod.tar
